Split distribution strategy tests
