+++ particula/particles/distribution_strategies/mass_based_moving_bin.py
@@ -1,0 +1,88 @@
+"""Mass-based moving bin strategy."""
+
+import logging
+import numpy as np
+from numpy.typing import NDArray
+
+from .base import DistributionStrategy
+
+logger = logging.getLogger("particula")
+
+
+class MassBasedMovingBin(DistributionStrategy):
+    """
+    Strategy for particles represented by their mass distribution.
+
+    Calculates particle mass, radius, and total mass based on the
+    particle mass, number concentration, and density. This moving-bin
+    approach adjusts mass bins on mass addition events.
+
+    Methods:
+    - get_name : Return the type of the distribution strategy.
+    - get_species_mass : Calculate the mass per species.
+    - get_mass : Calculate the mass of the particles or bin.
+    - get_total_mass : Calculate the total mass of particles.
+    - get_radius : Calculate the radius of particles.
+    - add_mass : Add mass to the particle distribution.
+    - add_concentration : Add concentration to the distribution.
+    """
+
+    def get_species_mass(
+        self, distribution: NDArray[np.float64], density: NDArray[np.float64]
+    ) -> NDArray[np.float64]:
+        return distribution
+
+    def get_radius(
+        self, distribution: NDArray[np.float64], density: NDArray[np.float64]
+    ) -> NDArray[np.float64]:
+        volumes = distribution / density
+        return (3 * volumes / (4 * np.pi)) ** (1 / 3)
+
+    def add_mass(
+        self,
+        distribution: NDArray[np.float64],
+        concentration: NDArray[np.float64],
+        density: NDArray[np.float64],
+        added_mass: NDArray[np.float64],
+    ) -> tuple[NDArray[np.float64], NDArray[np.float64]]:
+        return distribution + added_mass, concentration
+
+    def add_concentration(
+        self,
+        distribution: NDArray[np.float64],
+        concentration: NDArray[np.float64],
+        added_distribution: NDArray[np.float64],
+        added_concentration: NDArray[np.float64],
+    ) -> tuple[NDArray[np.float64], NDArray[np.float64]]:
+        if (distribution.shape != added_distribution.shape) and (
+            np.allclose(distribution, added_distribution, rtol=1e-6)
+        ):
+            message = (
+                "When adding concentration to MassBasedMovingBin, the distribution "
+                "and added distribution should match."
+            )
+            logger.error(message)
+            raise ValueError(message)
+        if concentration.shape != added_concentration.shape:
+            message = (
+                "When adding concentration to MassBasedMovingBin, the arrays "
+                "should have the same shape."
+            )
+            logger.error(message)
+            raise ValueError(message)
+
+        concentration += added_concentration
+        return distribution, concentration
+
+    def collide_pairs(
+        self,
+        distribution: NDArray[np.float64],
+        concentration: NDArray[np.float64],
+        density: NDArray[np.float64],
+        indices: NDArray[np.int64],
+    ) -> tuple[NDArray[np.float64], NDArray[np.float64]]:
+        message = (
+            "Colliding pairs in MassBasedMovingBin is not physically meaningful"
+        )
+        logger.warning(message)
+        raise NotImplementedError(message)
